Add files via upload
Given basic parameters of the RV-10 and RV-12iS aircrafts, I wrote a function and script in MATLAB to compare the performance capabilities of these two aircrafts. The focus of my project was on minimum drag, minimum drag speed, as well as flight range and a detailed analysis of operation costs.